Migrate from daScript to Lua(JIT)
This change was motivated by a desire to have a scripting language with a larger ecosystem so that if the engine was ever moved to another langauge entirely (e.g. Rust, Zig...jai?), there would be a higher likelihood for support (better known language with more libraries, easier bindings due to being C-based). Its performance is also more reputable.

Note that this includes the compiler for Teal, a Lua dialect which I'm going to try using as much as possible going forward. The Serpent library was also brought along.